+++ admin/tabs/cluster/connpool/ConnPoolMonitorPanel.ts
@@ -5,10 +5,10 @@
  * Copyright © 2024 Extremely Heavy Industries Inc.
  */
 import {AppModel} from '@xh/hoist/admin/AppModel';
-import {ConnPoolMonitorModel} from '@xh/hoist/admin/tabs/server/connectionpool/ConnPoolMonitorModel';
+import {ConnPoolMonitorModel} from '@xh/hoist/admin/tabs/cluster/connpool/ConnPoolMonitorModel';
 import {chart} from '@xh/hoist/cmp/chart';
 import {grid, gridCountLabel} from '@xh/hoist/cmp/grid';
-import {filler, hframe, vframe} from '@xh/hoist/cmp/layout';
+import {filler, hframe, span, vframe} from '@xh/hoist/cmp/layout';
 import {creates, hoistCmp} from '@xh/hoist/core';
 import {button, exportButton} from '@xh/hoist/desktop/cmp/button';
 import {errorMessage} from '@xh/hoist/desktop/cmp/error';
@@ -26,33 +26,28 @@ export const connPoolMonitorPanel = hoistCmp.factory({
             });
         }
 
-        if (model.lastLoadException) {
-            return errorMessage({
-                title: 'Error loading connection pool snapshots.',
-                error: model.lastLoadException,
-                actionFn: () => model.refreshAsync()
-            });
-        }
-
         const {readonly} = AppModel;
         return panel({
             tbar: [
+                span({
+                    item: 'JDBC Connection Pool',
+                    className: 'xh-bold'
+                }),
+                filler(),
+                gridCountLabel({unit: 'snapshot'}),
+                '-',
                 button({
                     text: 'Take Snapshot',
                     icon: Icon.camera(),
                     omit: readonly,
                     onClick: () => model.takeSnapshotAsync()
                 }),
-                '-',
                 button({
                     text: 'Reset Stats',
                     icon: Icon.reset(),
-                    intent: 'danger',
                     omit: readonly,
                     onClick: () => model.resetStatsAsync()
                 }),
-                filler(),
-                gridCountLabel({unit: 'snapshot'}),
                 '-',
                 exportButton()
             ],
@@ -77,20 +72,22 @@ export const connPoolMonitorPanel = hoistCmp.factory({
 const poolConfigPanel = hoistCmp.factory<ConnPoolMonitorModel>({
     render({model}) {
         return panel({
-            title: 'Connection Pool Configuration',
-            icon: Icon.gears(),
+            title: 'Pool Configuration',
+            icon: Icon.info(),
             compactHeader: true,
             modelConfig: {
-                defaultSize: 500,
-                defaultCollapsed: true,
-                side: 'right'
+                side: 'right',
+                defaultSize: 450,
+                defaultCollapsed: true
             },
             item: jsonInput({
-                value: JSON.stringify(model.poolConfiguration, null, 2),
                 readonly: true,
                 height: '100%',
                 width: '100%',
-                enableSearch: true
+                enableSearch: true,
+                showFullscreenButton: false,
+                editorProps: {lineNumbers: false},
+                value: JSON.stringify(model.poolConfiguration, null, 2)
             })
         });
     }